Add files via upload
Case Study A medical researcher obtains the systolic blood pressure readings (in mm Hg) in the accompanying list from a sample of women aged 18- 24 who have a new strain of viral infection. 134.9, 78.7, 108.9, 133.0, 123.7, 96.1, 126.9, 89.8, 132.0, 134.7, 132.1, 121.7, 112.3, 150.2, 158.3, 154.4 Healthy women in that age group have readings that are normally distributed with a mean blood pressure of 114.8.Test the claim that the sample comes from a healthy women population in that age group. [Data:blood_pressure.xlsx]